requirements.txt
langchain_openai==0.3.5
python-dotenv==1.0.1
